refactor: improve project routes organization
- Split project routes into smaller files by functionality:
  - management.ts: project CRUD operations
  - state.ts: project state management
  - chat.ts: AI chat functionality
  - git.ts: git operations
- Created index.ts to combine all routes
- Improved code organization and maintainability
